Move vendor templates to external files, replace Xur with component
