+++ src/scripts/vendors/vendor-items.html
@@ -1,0 +1,32 @@
+<div class="vendor-char-items"
+     ng-repeat="(vendorHash, vendor) in $ctrl.vendors | values | vendorTab:$ctrl.activeTab | orderBy:['vendorOrder'] track by vendor.hash">
+  <div class="title">
+    <div class="collapse-handle"
+         ng-click="$ctrl.toggleSection(vendorHash)">
+      <i class="fa collapse"
+         ng-class="$ctrl.settings.collapsedSections[vendorHash] ? 'fa-plus-square-o': 'fa-minus-square-o'"></i>
+      <img class="vendor-icon"
+           ng-src="{{::vendor.icon | bungieIcon}}" />
+      {{::vendor.name}}
+      <span class="vendor-location">{{::vendor.location}}</span>
+    </div>
+    <countdown class="vendor-timer"
+               ng-if="vendor.nextRefreshDate[0] !== '9'"
+               end-time="vendor.nextRefreshDate"></countdown>
+  </div>
+  <div class="vendor-row"
+       ng-if="!$ctrl.settings.collapsedSections[vendorHash]">
+    <vendor-currencies vendor-categories="vendor.categories"
+                       total-coins="$ctrl.totalCoins"
+                       property-filter="$ctrl.activeTab"></vendor-currencies>
+    <div class="vendor-category"
+         ng-repeat="category in vendor.categories | vendorTab:$ctrl.activeTab track by category.index">
+      <h3 class="category-title">{{::category.title}}</h3>
+      <div class="vendor-items">
+        <vendor-item ng-repeat="saleItem in category.saleItems | vendorTabItems:$ctrl.activeTab track by saleItem.index"
+                     sale-item="saleItem"
+                     total-coins="$ctrl.totalCoins"></vendor-item>
+      </div>
+    </div>
+  </div>
+</div>'
