+++ src/scripts/vendors/xur.component.js
@@ -1,0 +1,11 @@
+import template from './xur.html';
+
+export const Xur = {
+  template: template,
+  controller: XurController
+};
+
+function XurController(dimXurService) {
+  'ngInject';
+  this.xurService = dimXurService;
+}
