--- src/scripts/vendors/xur.controller.js
@@ -1,4 +1,0 @@
-export function XurController(dimXurService) {
-  'ngInject';
-  this.xurService = dimXurService;
-}
