+++ src/scripts/vendors/xur.html
@@ -1,8 +1,8 @@
-<div class="xurDialog" ng-controller="dimXurCtrl as vm">
+<div class="xurDialog">
   <h1 translate="Help.Xur"></h1>
   <div class="ngdialog-inner-content vendor">
     <div class="vendors">
-      <vendor-items vendors-data="vm.xurService.vendors" total-coins="vm.xurService.totalCoins" extra-move-popup-class="'xur-move-popup'"></vendor-items>
+      <vendor-items vendors-data="$ctrl.xurService.vendors" total-coins="$ctrl.xurService.totalCoins" extra-move-popup-class="'xur-move-popup'"></vendor-items>
     </div>
   </div>
 </div>
